refactor: have shared types for go
