+++ go/event-ingestion/kafka/producer.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/twmb/franz-go/pkg/kgo"
 	"github.com/uselotus/lotus/go/event-ingestion/config"
-	"github.com/uselotus/lotus/go/event-ingestion/types"
+	"github.com/uselotus/lotus/go/pkg/types"
 )
 
 type StreamEvent struct {
