+++ go/pkg/go.mod
@@ -1,0 +1,3 @@
+module github.com/uselotus/lotus/go/pkg
+
+go 1.20
